+++ merged-packages/eth-block-tracker/src/PollingBlockTracker.test.ts
@@ -1,13 +1,13 @@
 import { PollingBlockTracker } from '.';
-import buildDeferred from '../tests/buildDeferred';
-import EMPTY_FUNCTION from '../tests/emptyFunction';
-import recordCallsToSetTimeout from '../tests/recordCallsToSetTimeout';
-import { withPollingBlockTracker } from '../tests/withBlockTracker';
+import buildDeferred from '../test/buildDeferred';
+import EMPTY_FUNCTION from '../test/emptyFunction';
+import recordCallsToSetTimeout from '../test/recordCallsToSetTimeout';
+import { withPollingBlockTracker } from '../test/withBlockTracker';
 
-interface Sync {
+type Sync = {
   oldBlock: string;
   newBlock: string;
-}
+};
 
 const METHODS_TO_ADD_LISTENER = ['on', 'addListener'] as const;
 const METHODS_TO_REMOVE_LISTENER = ['off', 'removeListener'] as const;
@@ -174,7 +174,7 @@ describe('PollingBlockTracker', () => {
 
             await withPollingBlockTracker(async ({ blockTracker }) => {
               expect(blockTracker.isRunning()).toBe(false);
-              blockTracker.getLatestBlock();
+              await blockTracker.getLatestBlock();
               expect(blockTracker.isRunning()).toBe(false);
             });
           });
@@ -1203,7 +1203,7 @@ describe('PollingBlockTracker', () => {
           },
         },
         async ({ blockTracker }) => {
-          blockTracker.checkForLatestBlock();
+          await blockTracker.checkForLatestBlock();
           await new Promise((resolve) => {
             blockTracker.on('latest', resolve);
           });
@@ -1742,10 +1742,11 @@ describe('PollingBlockTracker', () => {
                 interceptCallback: (callback, stopPassingThroughCalls) => {
                   return async () => {
                     try {
-                      await callback();
+                      return await callback();
                     } catch (error: unknown) {
                       resolve(error);
                       stopPassingThroughCalls();
+                      return undefined;
                     }
                   };
                 },
@@ -2284,10 +2285,11 @@ describe('PollingBlockTracker', () => {
                 interceptCallback: (callback, stopPassingThroughCalls) => {
                   return async () => {
                     try {
-                      await callback();
+                      return await callback();
                     } catch (error: unknown) {
                       resolve(error);
                       stopPassingThroughCalls();
+                      return undefined;
                     }
                   };
                 },